+++ src/__test__/commentsCounter.test.js
@@ -1,5 +1,4 @@
 import commentsCounter from '../commentsCounter.js';
-import homeCounter from '../homecount.js'
 
 describe('returns the length of comments', () => {
   it('returns length of initial comments', () => {
@@ -21,24 +20,5 @@ describe('returns the length of comments', () => {
       },
     ];
     expect(commentsCounter(newComments)).toBe(2);
-  });
-});
-
-
-describe('Returns number of items on the homepage', () => {
-  it('returns 0 if the meals array is empty', () => {
-    const mealsArray = [];
-    expect(homeCounter(mealsArray)).toBe(0);
-  });
-
-  it('returns the number of meals in the mealsArray ', () => {
-    const mealsArray = [
-      {
-      strMeal:	"Canadian Butter Tarts",
-      strMealThumb:	"https://www.themealdb.com/images/media/meals/wpputp1511812960.jpg",
-      idMeal:	"52923"
-      }
-    ];
-    expect(homeCounter(mealsArray)).toBe(1);
   });
 });
